Consolidate .netcoreapp.cs files because System.Text.Encoding project is no longer cross-compiled
